parted: Display details of partition alignment failure (#726856)
When alignment for a new partition fails it isn't always obvious why it
failed. This adds printing the reason for the failure, in the form of:

start % grain != offset

This modifies align-check in interactive mode to print the alignment the
error details if it isn't aligned. Script mode behavior is unchanged.

Also cleanup pointer usage and handle asprintf failure by using a constant
string in the error report - "unknown (malloc failure)".