Use BLKSSZGET to get device sector size in _device_probe_geometry()
Seen on certain newer devices (such as >32G SDHC memory cards), the
HDIO_GETGEO ioctl does not return useful information.  The libparted
code records hardware and bios reported geometry information, but all of
that is largely unusable these days.  The information is used in the
PedConstraint code for aligning partitions.  The sector count is most
useful.  Rather than only trying HDIO_GETGIO, first initialize the
bios_geom fields to 0 and then use BLKSSZGET to capture the sector size.
If that fails, try HDIO_GETGEO.  And if that fails, raise a warning and
fall back on the library's default sector size macro.

This problem showed up on Raspberry Pi devices where users were
attempting to grow a partition to fill the SDHC card.  Using the
optimal_aligned_constraint returned invalid geometry information
(98703359 instead of 124735488 sectors).  The issue was reported here:

    https://github.com/fedberry/fedberry/issues/8

And to the pyparted project:

    https://github.com/rhinstaller/pyparted/issues/25

I've applied this patch locally to parted, rebuilt, and reinstalled it
and it is working correctly for the problem SDHC cards.

Signed-off-by: Brian C. Lane <bcl@redhat.com>