libparted: Don't warn if no HDIO_GET_IDENTITY ioctl
Fully virtualized Xen VMs (HVM) use Linux IDE devices which don't
support the HDIO_GET_IDENTITY ioctl. EINVAL is returned, a warning
is printed and the device model is set to "Generic IDE" in that
case. The problem is seeing the warning all the time. So drop it
for this case.

Signed-off-by: Brian C. Lane <bcl@redhat.com>