libparted: Fix MacOS boot support
boot_region_length (or BootSize in the MacOS dialect) is the length
of the driver code in the driver partition. This length is used
to compute the checksum of the driver.

libparted updates this value by setting the whole size of the partition
without computing the checksum of the driver using this size.

As the checksum is wrong, the driver is broken and cannot be loaded
by the MacOS ROM, and thus the disk is not bootable anymore.

Moreover, parted try to update the driver list and makes it disappear.

As parted is not able to insert a driver in a partition,
the driver is generally inserted by the Apple HD Tool,
this patch removes the line updating the driver size.

We also simplify the driver list scan and fix endianess use.

This has been tested... and it works, now.

I have updated a bootable disk with parted on x86_64 and
then been able to boot it (again) on a Mac LC III.

Signed-off-by: Laurent Vivier <laurent@vivier.eu>
Signed-off-by: Brian C. Lane <bcl@redhat.com>