libparted: Fix udev cookie leak in _dm_resize_partition
The function is setting udev cookies, but not using them when waiting
for the task. This results in leaked cookies, which can eventually
exhaust the available number of semaphores.

'dmsetup udevcookies' will show a cookie remaining afterwards, and 'ipcs -s'
will show the semaphores in use.

Also simplified the exit so that the task is always destroyed and memory
is all freed in the same path.

Resolves: rhbz#1455564
(cherry picked from commit e7870afe3c13dcc77845d48409daa35e3e42b5fb)