libparted/fs/amiga/affs.c: Remove printf() to avoid confusion
Currently, the affs code in libparted prints a debug line to stdout when
an Amiga FFS partition is probed. This confuses tools that parse stdout.

A prominent example is the partitioner (partman) in Debian's installer,
which parses stdout from a tool (parted_devices) that uses libparted to
scan the system's block devices. As of now, Debian cannot be installed
on a typical Amiga:

  https://lists.debian.org/debian-68k/2019/07/msg00015.html

This patch removes this unexpected printf().

Cc: glaubitz@debian.org
Signed-off-by: Max Staudt <max@enpas.org>
Signed-off-by: Brian C. Lane <bcl@redhat.com>