libparted: Fix a GCC warning -Wunused-but-set-variable
GCC warns that a variable 'prealloc' defined for _generic_affs_probe() in
fs/amiga/affs.c is set but its value is never used.

  CC       amiga/affs.lo
amiga/affs.c: In function '_generic_affs_probe':
amiga/affs.c:54:35: warning: variable 'prealloc' set but not used [-Wunused-but-set-variable]
   54 |  int blocksize = 1, reserved = 2, prealloc = 0;
      |                                   ^~~~~~~~
Remove the variable for simplicity and to avoid the warning.

Signed-off-by: Shin'ichiro Kawasaki <kawasaki@juno.dti.ne.jp>
Signed-off-by: Brian C. Lane <bcl@redhat.com>