parted: Preserve resizepart End when prompted for busy partition
Resizing busy partitions is allowed, but the user is prompted, which
erases the cmdline. It is annoying to have to re-end the ending location
after answering Yes. This saves the word and pushes it back onto the
cmdline after the user agrees to resize the busy partition.

(cherry picked from commit 691dabc930a6d544dfd8da787cddd159bab34b1e)

Resolves: rhbz#1861804