parted: add --fix to "fix" in script mode
Use-case: VMs are booted from images that are smaller than their virtual
disk. This means that -- almost by definition -- the secondary GPT
header will be "misplaced", i.e., not at the end of the virtual disk.

Without this patch, parted cannot be used for custom/exotic partitioning
when the VM boots (e.g., in cloud-init's `bootcmd`). Specifically, it
will fail as follows:

```
$ sudo parted --script /dev/vda "mkpart 2 10GB -1"
Warning: Not all of the space available to /dev/vda appears to be used,
you can fix the GPT to use all of the space (an extra 500 blocks) or continue with the current setting?
Error: Unable to satisfy all constraints on the partition.
```

This happens because, in script mode, exceptions are usually not
resolved.

This patch adds `--fix`. This allows exceptions to be automatically
resolved with Fix. As a result, the following command will work:

```
$ sudo parted --fix --script /dev/vda "mkpart 2 10GB -1"
```

Signed-off-by: Brian C. Lane <bcl@redhat.com>