Move Exception Option values into enum
Adding enums together doesn't create a new enum value, so when compiling
with warnings you will get warnings like:

warning: case value ‘96’ not in enumerated type

for PED_EXCEPTION_IGNORE_CANCEL

This moved the defines into the enum as new values so that they are
recognized as valid members of the enum with the values staying the
same.

NOTE: PED_EXCEPTION_OPTION_LAST *MUST* be the last of the individual
options, not the combined options.

Thanks to D. Hugh Redelmeier for this patch.