libparted: Add swap flag to msdos disklabel
Previously you had to set the filesystem type to one of the linux-swap
options at creation time. With this change you can now toggle the
partition swap type using the 'swap' partition flag in the same way that
you can on gpt disklabels.

Thanks to Arvin Schnell for this patch.