libparted: Fix problem with creating 1s partitions
There was a 1-off error in _partition_get_overlap_constraint that
prevented partitions from being created in 1s free space. You could
create 1s partitions as long they were done in order, but not after
leaving 'holes'.

This fixes this and adds tests for it on msdos and gpt disklabels.