parted: Add display of GPT UUIDs in JSON output
This adds 2 new disk type features, one for the whole disk UUID and
another for the per-partition UUID. It adds ped_disk_get_uuid and
ped_partition_get_uuid functions to retrieve them.

It adds them to the JSON output on GPT disklabeled disks as "uuid" in
the disk and partitions sections of the JSON output.

Signed-off-by: Brian C. Lane <bcl@redhat.com>