disk.in.h: Remove use of enums with #define
The preprocessor doesn't evaluate the enum, so it ends up being 0, which
causes problems for library users like pyparted which try to use the _LAST
value to conditionally include support for newer flags.

Instead just define the int that is the first and last entry in each enum.

Thanks to adamw and dcantrell for help arriving at a solution.