docs: Finish setup of libparted API docs
I think 16 years is long enough to wait. Back in 2007 the doc/API file
was being transitioned to a doxygen document at
https://www.gnu.org/software/parted/api/ but it was never finished, and
never updated.

This commit updates it to be compatible with Doxygen 1.9.6, and it
updates doc/API to be doc/API.md which is used as the main page of the
document. I'm sure this isn't complete, but it gives a good starting
point for further maintenance. Updating the webpage has been documented
at the end of the README-release file.