libparted: Fix sun disklabel unhandled exception
The CHS warning should only continue if ignored, not if unhandled.

Script mode, or exception handlers can return PED_EXCEPTION_UNHANDLED
which should act the same as a cancel. Previously it would only exit if
cancel was selected, allowing it to continue to use the bad CHS and
crash later.